Schema -> CodeGeneratingSchema; pull out interface
