+++ atlasdb-cassandra-tests/src/test/java/com/palantir/atlasdb/keyvalue/cassandra/CQLKeyValueService.java
@@ -97,6 +97,7 @@ import com.palantir.atlasdb.keyvalue.api.Value;
 import com.palantir.atlasdb.keyvalue.impl.AbstractKeyValueService;
 import com.palantir.atlasdb.keyvalue.impl.Cells;
 import com.palantir.atlasdb.keyvalue.impl.KeyValueServices;
+import com.palantir.atlasdb.table.description.DefaultTableMetadata;
 import com.palantir.atlasdb.table.description.TableMetadata;
 import com.palantir.atlasdb.transaction.impl.TransactionConstants;
 import com.palantir.common.annotation.Idempotent;
@@ -1179,7 +1180,7 @@ public class CQLKeyValueService extends AbstractKeyValueService {
         double falsePositiveChance = CassandraConstants.DEFAULT_LEVELED_COMPACTION_BLOOM_FILTER_FP_CHANCE;
 
         if (rawMetadata != null && rawMetadata.length != 0) {
-            TableMetadata tableMetadata = TableMetadata.BYTES_HYDRATOR.hydrateFromBytes(rawMetadata);
+            TableMetadata tableMetadata = DefaultTableMetadata.BYTES_HYDRATOR.hydrateFromBytes(rawMetadata);
             dbCompressionRequested = tableMetadata.isDbCompressionRequested();
             rangeScanAllowed = tableMetadata.isRangeScanAllowed();
             negativeLookups = tableMetadata.hasNegativeLookups();
