+++ atlasdb-client/src/main/java/com/palantir/atlasdb/table/description/IndexMetadata.java
@@ -15,275 +15,11 @@
  */
 package com.palantir.atlasdb.table.description;
 
-import java.util.List;
+public interface IndexMetadata {
 
-import javax.annotation.Nullable;
+	/**
+	 * @return the table metadata for the index table
+	 */
+	TableMetadata getTableMetadata();
 
-import org.apache.commons.lang.Validate;
-
-import com.google.common.collect.ImmutableList;
-import com.google.common.collect.Iterables;
-import com.google.common.collect.Lists;
-import com.palantir.atlasdb.protos.generated.TableMetadataPersistence.CachePriority;
-import com.palantir.atlasdb.protos.generated.TableMetadataPersistence.ExpirationStrategy;
-import com.palantir.atlasdb.protos.generated.TableMetadataPersistence.PartitionStrategy;
-import com.palantir.atlasdb.protos.generated.TableMetadataPersistence.SweepStrategy;
-import com.palantir.atlasdb.table.description.IndexDefinition.IndexType;
-import com.palantir.atlasdb.table.description.render.Renderers;
-import com.palantir.atlasdb.transaction.api.ConflictHandler;
-
-public class IndexMetadata {
-    final String name;
-    final String javaName;
-    final ImmutableList<IndexComponent> rowComponents;
-    final ImmutableList<IndexComponent> colComponents;
-    @Nullable final String columnNameToGetData;
-    final CachePriority cachePriority;
-    final PartitionStrategy paritionStrategy;
-    boolean rangeScanAllowed;
-    boolean dbCompressionRequested;
-    boolean negativeLookups;
-    final ConflictHandler conflictHandler;
-    final IndexCondition indexCondition;
-    IndexType indexType;
-    final SweepStrategy sweepStrategy;
-    private final ExpirationStrategy expirationStrategy;
-
-    public static IndexMetadata createIndex(String name,
-                                                    String javaName,
-                                                    Iterable<IndexComponent> rowComponents,
-                                                    CachePriority cachePriority,
-                                                    PartitionStrategy paritionStrategy,
-                                                    ConflictHandler conflictHandler,
-                                                    boolean rangeScanAllowed,
-                                                    boolean dbCompressionRequested,
-                                                    boolean negativeLookups,
-                                                    IndexCondition indexCondition,
-                                                    IndexType indexType,
-                                                    SweepStrategy sweepStrategy,
-                                                    ExpirationStrategy expirationStrategy) {
-        Validate.isTrue(!Iterables.isEmpty(rowComponents));
-        Iterable<IndexComponent> colComponents = ImmutableList.<IndexComponent>of();
-        return new IndexMetadata(
-                name,
-                javaName,
-                rowComponents,
-                colComponents,
-                getColNameToAccessFrom(rowComponents, colComponents, indexCondition),
-                cachePriority,
-                paritionStrategy,
-                conflictHandler,
-                rangeScanAllowed,
-                dbCompressionRequested,
-                negativeLookups,
-                indexCondition,
-                indexType,
-                sweepStrategy,
-                expirationStrategy);
-    }
-
-    public static IndexMetadata createDynamicIndex(String name,
-                                                           String javaName,
-                                                           Iterable<IndexComponent> rowComponents,
-                                                           Iterable<IndexComponent> colComponents,
-                                                           CachePriority cachePriority,
-                                                           PartitionStrategy paritionStrategy,
-                                                           ConflictHandler conflictHandler,
-                                                           boolean rangeScanAllowed,
-                                                           boolean dbCompressionRequested,
-                                                           boolean negativeLookups,
-                                                           IndexCondition indexCondition,
-                                                           IndexType indexType,
-                                                           SweepStrategy sweepStrategy,
-                                                           ExpirationStrategy expirationStrategy) {
-        Validate.isTrue(!Iterables.isEmpty(rowComponents));
-        Validate.isTrue(!Iterables.isEmpty(colComponents));
-        return new IndexMetadata(
-                name,
-                javaName,
-                rowComponents,
-                colComponents,
-                getColNameToAccessFrom(rowComponents, colComponents, indexCondition),
-                cachePriority,
-                paritionStrategy,
-                conflictHandler,
-                rangeScanAllowed,
-                dbCompressionRequested,
-                negativeLookups,
-                indexCondition,
-                indexType,
-                sweepStrategy,
-                expirationStrategy);
-    }
-
-    private IndexMetadata(String name,
-                          String javaName,
-                          Iterable<IndexComponent> rowComponents,
-                          Iterable<IndexComponent> colComponents,
-                          String colNameToAccessFrom,
-                          CachePriority cachePriority,
-                          PartitionStrategy partitionStrategy,
-                          ConflictHandler conflictHandler,
-                          boolean rangeScanAllowed,
-                          boolean dbCompressionRequested,
-                          boolean negativeLookups,
-                          IndexCondition indexCondition,
-                          IndexType indexType,
-                          SweepStrategy sweepStrategy,
-                          ExpirationStrategy expirationStrategy) {
-        this.name = name;
-        this.javaName = javaName;
-        this.rowComponents = ImmutableList.copyOf(rowComponents);
-        this.colComponents = ImmutableList.copyOf(colComponents);
-        this.columnNameToGetData = colNameToAccessFrom;
-        this.cachePriority = cachePriority;
-        this.paritionStrategy = partitionStrategy;
-        this.conflictHandler = conflictHandler;
-        this.rangeScanAllowed = rangeScanAllowed;
-        this.dbCompressionRequested = dbCompressionRequested;
-        this.negativeLookups = negativeLookups;
-        this.indexCondition = indexCondition;
-        this.indexType = indexType;
-        this.sweepStrategy = sweepStrategy;
-        this.expirationStrategy = expirationStrategy;
-    }
-
-    private static String getColNameToAccessFrom(Iterable<IndexComponent> rowComponents,
-                                          Iterable<IndexComponent> colComponents,
-                                          IndexCondition indexCondition) {
-
-        String colNameToAccessFrom = null;
-        for (IndexComponent indexComponent : Iterables.concat(rowComponents, colComponents)) {
-            if (indexComponent.columnNameToGetData == null) {
-                continue;
-            }
-            if (colNameToAccessFrom == null) {
-                colNameToAccessFrom = indexComponent.columnNameToGetData;
-            }
-            if (!indexComponent.columnNameToGetData.equals(colNameToAccessFrom)) {
-                throw new IllegalArgumentException("An index must only reference one column."
-                        + "This references " + indexComponent.columnNameToGetData + " and "
-                        + colNameToAccessFrom);
-            }
-        }
-        if (indexCondition != null) {
-            if (colNameToAccessFrom == null) {
-                colNameToAccessFrom = indexCondition.getSourceColumn();
-            }
-            if (!indexCondition.getSourceColumn().equals(colNameToAccessFrom)) {
-                throw new IllegalArgumentException("An index must only reference one column."
-                        + "This references " + indexCondition.getSourceColumn() + " and "
-                        + colNameToAccessFrom);
-            }
-        }
-        return colNameToAccessFrom;
-    }
-
-    public TableMetadata getTableMetadata() {
-        List<NameComponentDescription> rowDescList = Lists.newArrayList();
-        for (IndexComponent indexComp : rowComponents) {
-            rowDescList.add(indexComp.rowKeyDesc);
-        }
-
-        ColumnMetadataDescription column;
-        if (indexType.equals(IndexType.ADDITIVE)) {
-            if (colComponents.isEmpty()) {
-                column = getAdditiveIndexColumn();
-            } else {
-                List<NameComponentDescription> colDescList = Lists.newArrayList();
-                for (IndexComponent indexComp : colComponents) {
-                    colDescList.add(indexComp.rowKeyDesc);
-                }
-                column = getDynamicAdditiveIndexColumn(colDescList);
-            }
-        } else if (indexType.equals(IndexType.CELL_REFERENCING)) {
-            List<NameComponentDescription> colDescList = Lists.newArrayList();
-            for (IndexComponent indexComp : colComponents) {
-                colDescList.add(indexComp.rowKeyDesc);
-            }
-            column = getCellReferencingIndexColumn(colDescList);
-        } else {
-            throw new IllegalArgumentException("Unknown index type " + indexType);
-        }
-        return new TableMetadata(
-                new NameMetadataDescription(rowDescList),
-                column,
-                conflictHandler,
-                cachePriority,
-                paritionStrategy,
-                rangeScanAllowed,
-                dbCompressionRequested,
-                negativeLookups,
-                sweepStrategy,
-                expirationStrategy);
-    }
-
-    public boolean isDynamicIndex() {
-        return !colComponents.isEmpty();
-    }
-
-    public List<IndexComponent> getRowComponents() {
-        return rowComponents;
-    }
-
-    /**
-     * Calling this is only meaningful if {@link #isDynamicIndex()} returns true.
-     */
-    public List<IndexComponent> getColumnComponents() {
-        return colComponents;
-    }
-
-    @Nullable
-    public String getColumnNameToAccessData() {
-        return columnNameToGetData;
-    }
-
-    public String getIndexName() {
-        return name;
-    }
-
-    public String getJavaIndexName() {
-        return javaName;
-    }
-
-    public IndexCondition getIndexCondition() {
-        return indexCondition;
-    }
-
-    public IndexType getIndexType() {
-        return indexType;
-    }
-
-    public String getIndexTable() {
-        String indexTable = getJavaIndexName();
-        if (indexTable == null) {
-            indexTable = Renderers.CamelCase(getIndexName());
-        }
-        return indexTable;
-    }
-
-    private static ColumnMetadataDescription getAdditiveIndexColumn() {
-        ColumnValueDescription columnValue = ColumnValueDescription.forType(ValueType.VAR_LONG);
-        NamedColumnDescription namedColumn = new NamedColumnDescription("e", "exists", columnValue);
-        return new ColumnMetadataDescription(ImmutableList.of(namedColumn));
-    }
-
-    private static ColumnMetadataDescription getDynamicAdditiveIndexColumn(List<NameComponentDescription> components) {
-        NameMetadataDescription columnDescription = new NameMetadataDescription(components);
-        ColumnValueDescription columnValue = ColumnValueDescription.forType(ValueType.VAR_LONG);
-        DynamicColumnDescription dynamicColumn = new DynamicColumnDescription(columnDescription, columnValue);
-        return new ColumnMetadataDescription(dynamicColumn);
-    }
-
-    private static ColumnMetadataDescription getCellReferencingIndexColumn(List<NameComponentDescription> components) {
-        components = ImmutableList.<NameComponentDescription>builder()
-                .add(new NameComponentDescription("row_name", ValueType.SIZED_BLOB))
-                .add(new NameComponentDescription("column_name", ValueType.SIZED_BLOB))
-                .addAll(components)
-                .build();
-        NameMetadataDescription columnDescription = new NameMetadataDescription(components);
-        ColumnValueDescription columnValue = ColumnValueDescription.forType(ValueType.VAR_LONG);
-        DynamicColumnDescription dynamicColumn = new DynamicColumnDescription(columnDescription, columnValue);
-        return new ColumnMetadataDescription(dynamicColumn);
-    }
 }
