+++ atlasdb-client/src/main/java/com/palantir/atlasdb/table/description/TableDefinitions.java
@@ -19,18 +19,18 @@ import com.google.common.collect.ImmutableList;
 import com.palantir.atlasdb.transaction.api.ConflictHandler;
 
 public class TableDefinitions {
-    private static final TableMetadata RAW_METADATA = new TableMetadata(
+    private static final TableMetadata RAW_METADATA = new DefaultTableMetadata(
             new NameMetadataDescription(ImmutableList.of(new NameComponentDescription("row", ValueType.BLOB))),
             new ColumnMetadataDescription(new DynamicColumnDescription(new NameMetadataDescription(ImmutableList.of(new NameComponentDescription("row", ValueType.BLOB))), ColumnValueDescription.forType(ValueType.BLOB))),
             ConflictHandler.SERIALIZABLE);
 
-    private static final TableMetadata JSON_METADATA = new TableMetadata(
+    private static final TableMetadata JSON_METADATA = new DefaultTableMetadata(
             new NameMetadataDescription(ImmutableList.of(new NameComponentDescription("row", ValueType.STRING))),
             new ColumnMetadataDescription(new DynamicColumnDescription(new NameMetadataDescription(ImmutableList.of(new NameComponentDescription("c", ValueType.STRING))), ColumnValueDescription.forType(ValueType.BLOB))),
             ConflictHandler.SERIALIZABLE);
 
     private static final TableMetadata singleColumn(String colName) {
-        return new TableMetadata(
+        return new DefaultTableMetadata(
                 new NameMetadataDescription(ImmutableList.of(new NameComponentDescription("row", ValueType.STRING))),
                 new ColumnMetadataDescription(new DynamicColumnDescription(new NameMetadataDescription(ImmutableList.of(new NameComponentDescription(colName, ValueType.STRING))), ColumnValueDescription.forType(ValueType.BLOB))),
                 ConflictHandler.SERIALIZABLE);
