+++ atlasdb-client/src/main/java/com/palantir/atlasdb/table/description/TableMetadata.java
@@ -15,284 +15,25 @@
  */
 package com.palantir.atlasdb.table.description;
 
-import javax.annotation.concurrent.Immutable;
-
-import com.google.common.base.Preconditions;
-import com.google.protobuf.InvalidProtocolBufferException;
-import com.palantir.atlasdb.protos.generated.TableMetadataPersistence;
-import com.palantir.atlasdb.protos.generated.TableMetadataPersistence.CachePriority;
-import com.palantir.atlasdb.protos.generated.TableMetadataPersistence.ExpirationStrategy;
-import com.palantir.atlasdb.protos.generated.TableMetadataPersistence.PartitionStrategy;
 import com.palantir.atlasdb.protos.generated.TableMetadataPersistence.SweepStrategy;
-import com.palantir.atlasdb.protos.generated.TableMetadataPersistence.TableMetadata.Builder;
 import com.palantir.atlasdb.transaction.api.ConflictHandler;
-import com.palantir.common.base.Throwables;
-import com.palantir.common.persist.Persistable;
 
-@Immutable
-public class TableMetadata implements Persistable {
-    final NameMetadataDescription rowMetadata;
-    final ColumnMetadataDescription columns;
-    final ConflictHandler conflictHandler;
-    final CachePriority cachePriority;
-    final PartitionStrategy partitionStrategy;
-    final boolean rangeScanAllowed;
-    final boolean dbCompressionRequested;
-    final boolean negativeLookups;
-    final SweepStrategy sweepStrategy;
-    final ExpirationStrategy expirationStrategy;
+public interface TableMetadata {
 
-    public TableMetadata() {
-        this(
-                new NameMetadataDescription(),
-                new ColumnMetadataDescription(),
-                ConflictHandler.RETRY_ON_WRITE_WRITE);
-    }
+	NameMetadataDescription getRowMetadata();
 
-    public TableMetadata(NameMetadataDescription rowMetadata,
-                         ColumnMetadataDescription columns,
-                         ConflictHandler conflictHandler) {
-        this(
-                rowMetadata,
-                columns,
-                conflictHandler,
-                CachePriority.WARM,
-                PartitionStrategy.ORDERED,
-                false,
-                false,
-                false,
-                SweepStrategy.CONSERVATIVE,
-                ExpirationStrategy.NEVER);
-    }
+	ColumnMetadataDescription getColumns();
 
-    public TableMetadata(NameMetadataDescription rowMetadata,
-                         ColumnMetadataDescription columns,
-                         ConflictHandler conflictHandler,
-                         CachePriority cachePriority,
-                         PartitionStrategy partitionStrategy,
-                         boolean rangeScanAllowed,
-                         boolean dbCompressionRequested,
-                         boolean negativeLookups,
-                         SweepStrategy sweepStrategy,
-                         ExpirationStrategy expirationStrategy) {
-        if (rangeScanAllowed) {
-            Preconditions.checkArgument(
-                    partitionStrategy == PartitionStrategy.ORDERED,
-                    "range scan is only allowed if partition strategy is ordered");
-        }
-        this.rowMetadata = rowMetadata;
-        this.columns = columns;
-        this.conflictHandler = conflictHandler;
-        this.cachePriority = cachePriority;
-        this.partitionStrategy = partitionStrategy;
-        this.rangeScanAllowed = rangeScanAllowed;
-        this.dbCompressionRequested = dbCompressionRequested;
-        this.negativeLookups = negativeLookups;
-        this.sweepStrategy = sweepStrategy;
-        this.expirationStrategy = expirationStrategy;
-    }
+	boolean isRangeScanAllowed();
 
-    public NameMetadataDescription getRowMetadata() {
-        return rowMetadata;
-    }
+	boolean isDbCompressionRequested();
 
-    public ColumnMetadataDescription getColumns() {
-        return columns;
-    }
+	boolean hasNegativeLookups();
 
-    public boolean isWriteWriteConflict() {
-        return conflictHandler != ConflictHandler.IGNORE_ALL;
-    }
+	byte[] persistToBytes();
 
-    public ConflictHandler getConflictHandler() {
-        return conflictHandler;
-    }
+	ConflictHandler getConflictHandler();
 
-    public CachePriority getCachePriority() {
-        return cachePriority;
-    }
-
-    public PartitionStrategy getPartitionStrategy() {
-        return partitionStrategy;
-    }
-
-    public boolean isRangeScanAllowed() {
-        return rangeScanAllowed;
-    }
-
-    public boolean isDbCompressionRequested() {
-        return dbCompressionRequested;
-    }
-
-    public boolean hasNegativeLookups() {
-        return negativeLookups;
-    }
-
-    public SweepStrategy getSweepStrategy() {
-        return sweepStrategy;
-    }
-
-    public ExpirationStrategy getExpirationStrategy() {
-        return expirationStrategy;
-    }
-
-    @Override
-    public byte[] persistToBytes() {
-        return persistToProto().build().toByteArray();
-    }
-
-    public static final Hydrator<TableMetadata> BYTES_HYDRATOR = new Hydrator<TableMetadata>() {
-        @Override
-        public TableMetadata hydrateFromBytes(byte[] input) {
-            try {
-                TableMetadataPersistence.TableMetadata message = TableMetadataPersistence.TableMetadata.parseFrom(input);
-                return hydrateFromProto(message);
-            } catch (InvalidProtocolBufferException e) {
-                throw Throwables.throwUncheckedException(e);
-            }
-        }
-    };
-
-    public TableMetadataPersistence.TableMetadata.Builder persistToProto() {
-        Builder builder = TableMetadataPersistence.TableMetadata.newBuilder();
-        builder.setConflictHandler(ConflictHandlers.persistToProto(getConflictHandler()));
-        builder.setRowName(rowMetadata.persistToProto());
-        builder.setColumns(columns.persistToProto());
-        builder.setCachePriority(cachePriority);
-        builder.setPartitionStrategy(partitionStrategy);
-        builder.setRangeScanAllowed(rangeScanAllowed);
-        builder.setDbCompressionRequested(dbCompressionRequested);
-        builder.setNegativeLookups(negativeLookups);
-        builder.setSweepStrategy(sweepStrategy);
-        // expiration strategy doesn't need to be persisted.
-        return builder;
-    }
-
-    public static TableMetadata hydrateFromProto(TableMetadataPersistence.TableMetadata message) {
-        CachePriority cachePriority = CachePriority.WARM;
-        if (message.hasCachePriority()) {
-            cachePriority = message.getCachePriority();
-        }
-        PartitionStrategy partitionStrategy = PartitionStrategy.ORDERED;
-        if (message.hasPartitionStrategy()) {
-            partitionStrategy = message.getPartitionStrategy();
-        }
-        boolean rangeScanAllowed = false;
-        if (message.hasRangeScanAllowed()) {
-            rangeScanAllowed = message.getRangeScanAllowed();
-        }
-        boolean dbCompressionRequested = false;
-        if (message.hasDbCompressionRequested()) {
-            dbCompressionRequested = message.getDbCompressionRequested();
-        }
-        boolean negativeLookups = false;
-        if (message.hasNegativeLookups()) {
-            negativeLookups = message.getNegativeLookups();
-        }
-        SweepStrategy sweepStrategy = SweepStrategy.CONSERVATIVE;
-        if (message.hasSweepStrategy()) {
-            sweepStrategy = message.getSweepStrategy();
-        }
-        return new TableMetadata(
-                NameMetadataDescription.hydrateFromProto(message.getRowName()),
-                ColumnMetadataDescription.hydrateFromProto(message.getColumns()),
-                ConflictHandlers.hydrateFromProto(message.getConflictHandler()),
-                cachePriority,
-                partitionStrategy,
-                rangeScanAllowed,
-                dbCompressionRequested,
-                negativeLookups,
-                sweepStrategy,
-                ExpirationStrategy.NEVER);
-    }
-
-    @Override
-    public String toString() {
-        return "TableMetadata ["
-                + "rowMetadata=" + rowMetadata
-                + ", columns=" + columns
-                + ", conflictHandler=" + conflictHandler
-                + ", partitionStrategy=" + partitionStrategy
-                + ", rowMetadata =" + rowMetadata
-                + ", rangeScanAllowed =" + rangeScanAllowed
-                + ", dbCompressionRequested =" + dbCompressionRequested
-                + ", negativeLookups = " + negativeLookups
-                + ", sweepStrategy = " + sweepStrategy
-                + "]";
-    }
-
-    @Override
-    public int hashCode() {
-        final int prime = 31;
-        int result = 1;
-        result = prime * result + ((cachePriority == null) ? 0 : cachePriority.hashCode());
-        result = prime * result + ((columns == null) ? 0 : columns.hashCode());
-        result = prime * result + ((conflictHandler == null) ? 0 : conflictHandler.hashCode());
-        result = prime * result + (dbCompressionRequested ? 1231 : 1237);
-        result = prime * result + ((partitionStrategy == null) ? 0 : partitionStrategy.hashCode());
-        result = prime * result + (rangeScanAllowed ? 1231 : 1237);
-        result = prime * result + ((rowMetadata == null) ? 0 : rowMetadata.hashCode());
-        result = prime * result + (rangeScanAllowed? 0 : 1);
-        result = prime * result + (dbCompressionRequested? 0 : 1);
-        result = prime * result + (negativeLookups? 0 : 1);
-        result = prime * result + (sweepStrategy.hashCode());
-        return result;
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj) {
-            return true;
-        }
-        if (obj == null) {
-            return false;
-        }
-        if (getClass() != obj.getClass()) {
-            return false;
-        }
-        TableMetadata other = (TableMetadata) obj;
-        if (cachePriority != other.cachePriority) {
-            return false;
-        }
-        if (columns == null) {
-            if (other.columns != null) {
-                return false;
-            }
-        } else if (!columns.equals(other.columns)) {
-            return false;
-        }
-        if (conflictHandler != other.conflictHandler) {
-            return false;
-        }
-        if (dbCompressionRequested != other.dbCompressionRequested) {
-            return false;
-        }
-        if (partitionStrategy != other.partitionStrategy) {
-            return false;
-        }
-        if (rangeScanAllowed != other.rangeScanAllowed) {
-            return false;
-        }
-        if (rowMetadata == null) {
-            if (other.rowMetadata != null) {
-                return false;
-            }
-        } else if (!rowMetadata.equals(other.rowMetadata)) {
-            return false;
-        }
-        if (rangeScanAllowed != other.rangeScanAllowed) {
-            return false;
-        }
-        if (dbCompressionRequested != other.dbCompressionRequested) {
-            return false;
-        }
-        if (negativeLookups != other.negativeLookups) {
-            return false;
-        }
-        if (sweepStrategy != other.sweepStrategy) {
-            return false;
-        }
-        return true;
-    }
+	SweepStrategy getSweepStrategy();
 
 }
